client/requirements.txt
cryptography==3.4.7
pycryptodome==3.18.0
PyQt5==5.15.9
PyQt5_sip==12.11.0
